pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dataframe-analyst-mcp"
version = "0.1.0"
description = "MCP server for CSV/Excel/Google Sheets/Drive EDA"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name", email = "you@example.com" }]
dependencies = [
  "mcp>=1.14.0",
  "pandas>=2.2",
  "openpyxl>=3.1",
  "xlrd==1.2.0",
  "pydantic>=2.6"
]

[project.optional-dependencies]
gdrive = [
  "gspread>=6",
  "google-auth>=2.28",
  "google-auth-oauthlib>=1.2",
  "pydrive2>=1.19"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
